backEnd/charging-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pymongo==4.6.0
redis==5.0.1
websockets==12.0
ocpp==0.15.0
python-dotenv==1.0.0
httpx==0.25.2